Moved TTA into its own file
